make progress bar and adjust the card as well
