+++ components/summaries/content-section.tsx
@@ -1,0 +1,3 @@
+export default function ContentSection(){
+  return <div>content Section</div>
+}
