+++ components/summaries/summary-viewer.tsx
@@ -1,35 +1,15 @@
 "use client";
 import { useState } from "react";
-import { Card, CardContent, CardHeader, CardTitle } from "../ui/card";
+import { Card } from "../ui/card";
 import {NavigationControls} from "./navigation-controls";
-
-const parseSection = (section: string) => {
-  const [title, ...content] = section.split("\n");
-  const cleanTitle = title.startsWith("#") ? title.substring(1).trim() : title.trim();
-  
+import ProgressBar from "./progress-bar";
+import { parseSection } from "@/utils/summary-helper";
 
 
-  const points: string[] = [];
-  let currentPoint = "";
-
-  content.forEach((line) => {
-    const trimmedLine = line.trim();
-    if (trimmedLine.startsWith("-")) {
-      if (currentPoint) {
-        points.push(currentPoint);
-        currentPoint = "";
-      }
-      currentPoint = trimmedLine.substring(2).trim();
-    } else {
-      currentPoint += " " + trimmedLine;
-    }
-  });
-
-  if (currentPoint) {
-    points.push(currentPoint);
-  }
-
-  return { title: cleanTitle, content: points.filter((point)=>point && !point.startsWith("#") && !point.startsWith("[Choose")) };
+const SectionTitle = ({title}:{title: string}) => {
+  return <div  className="flex flex-col gap-2 mb-6 sticky top-0 pt-2 pb-4 bg-backgroun/80 backdrop-blur-xs z-10 ">
+    <h2 className="text-3xl lg-text-4xl font-bold text-center flex items-center justify-centergap-2">{title}</h2>
+  </div>
 };
 
 export default function SummaryViewer({ summary }: { summary: string }) {
@@ -50,10 +30,10 @@ const handlePrevious = () => {
 };
   return (
     <Card className="relative px-2 h-[500px] sm:h-[600px] lg:h-[700px] w-full xl:w-[600px] overflow-hidden bg-linear-to-br from-background via-background/95 to-rose-500/5 backdrop:blur-lg shadow-2xl rounded-3xl border border-rose-500/10"> 
-      
+      <ProgressBar section={sections} currentSection={currentSection} />
         <div className="h-full overflow-y-auto scrollbar-hide pt-12 sm:pt-16 pb-20 sm:pb-24">
           <div className="px-4 sm:px-6">
-            <h2>{sections[currentSection]?.title || ""}</h2>
+            <SectionTitle title={sections[currentSection]?.title || ""} />
             <ul>
               {sections[currentSection]?.content.map((point, index) => (
                 <li key={index}>{point}</li>
